Pass player position to Chunk AddPlayer()
This way the client can spawn the player on it's last known position instead of Vector2.zero

+++ Backend.Orleans.SharedContracts/Serialization/SerializableVector2.cs
@@ -2,11 +2,11 @@ namespace Backend.Orleans.SharedContracts.Serialization;
 
 [GenerateSerializer]
 public class SerializableVector2 {
-    [Id(0)] public float X { get; set; }
+    [Id(0)] public int X { get; set; }
 
-    [Id(1)] public float Y { get; set; }
+    [Id(1)] public int Y { get; set; }
 
-    public SerializableVector2(float x, float y) {
+    public SerializableVector2(int x, int y) {
         X = x;
         Y = y;
     }
